eglibc stage2 should not disable multilib
Adam Conrad requested that the default eglibc builds should be multilib
(like gcc builds are). Rather there should be some toggle to disable
multilib. That toggle should be a build profile. Since there currently
is no profile for requesting to disable multilib, I decided that the
profile will be named "nobiarch" (in resemblance of DEB_CROSS_NO_BIARCH)
for now.