also enable gcc-base in control.m4
It is needlessly conditional on TARGET being empty, which covers all the
cross cases. But it is also conditional on "ifenabled gccbase" which is
never the case for cross. So just drop the TARGET check to enable it
precisely when needed. The only change happens for
with_deps_on_target_arch_pkgs=yes.