fix gcc stage3 for x32
Installation failed like this:

    mv debian/tmp/usr/lib/x86_64-linux-gnux32/libvtv*.a debian/libgcc-4.9-dev/usr/lib/gcc/x86_64-linux-gnux32/4.9/
    mv: cannot stat 'debian/tmp/usr/lib/x86_64-linux-gnux32/libvtv*.a': No such file or directory

This is due to the subdirectories src/libvtv and src/libcilkrts being
added in gcc-4.9 and cross-ma-install-location.diff not being adapted to
the newer gcc version (removal of mudflap was an earlier problem). The
changes needed for these subdirectories are analogous to those needed
for existing subdirectories.