fix glibc's local-bootstrap-headers.diff
This patch attempts to create (empty) stubs.h, because its presence is
needed but it is usually derived from the built objects. The patch
completely missed out on the stubs-$abi.h parts and thus makes the
extra_install_ functions of the Debian glibc build fail. The updated
does not change the generation of stubs.h, but rather creates empty
stubs-$abi.h parts.