pass the architecture to builddep_*
The original first parameters was the profiles, but no function actually
used it. This inserts the host architecture as a new first parameter for
builddep_* (as already used by builddep_dpkg). It should allow
builddep_* to be reused for both native and cross builds which may be
necessary for e.g. libtool.