add stage1 to util-linux and fix build
util-linux got uploaded to sid. It now Build-Depends on pam and systemd,
has a cycle with systemd and generally breaks cross building. Add a
stage1 for avoiding the extra dependencies (thanks to Andreas
Henriksson), and fix the remaining issues.

It still uses the old profile syntax, because sid's dpkg is broken.