glibc: pass -l... to dh_shlibdeps
dpkg_shlibdeps has some hacks for multilib. Unfortunately, these do not
cover libx32 and libo32 (mips32). Therefore, the multilib builds for
i386 and mips64el fail at the dpkg_shlibdeps invocation failing to find
libc.so.

In the relevant make target, $(curpass) refers to a binary package name.
We assume that the part after the last dash is a glibc pass. A pass name
can be used to look up the slibdir which contains the libc. The -l flag
is only passed, if a slibdir was actually specified.