convert the /usr/<triplet>/include path hack to a glibc patch
The glibc package currently expects linux-libc-dev headers in the path
created by dpkg-cross, i.e. /usr/<triplet>/include. Thus we were
symlinking this to the linux-libc-dev:$ARCH path. Instead of this hack,
teach glibc where it actually lives.

Thanks to Dima Kogan for trying to reproduce the issues caused by the
hack.

Thanks to YunQiang Su for providing the patch.