allow comparing cross-built packages against native ones
This feature is enabled using the ENABLE_DEBBINDIFF=yes environment
variable. It only works for release architectures and only when the host
architecture is in multiarch-sync with the build architecture. Then it
downloads the natively built package from the archive and compares it to
the cross-built one using debbindiff from the reproducible toolchain.