start automatically ordering builds using dose
In addition to manually determining a build order of all required
packages, there now is a variable "need_packages" containing packages
with no specific order. These packages are tried when the
automatically_cross_build_packages function is invoked as soon as their
build-depends become satisfiable as determined by dose-builddebcheck.
Thus a manual build order can now be interleaved with phases of
automatic ordering.

Many thanks to Johannes Schauer for his help with this.