extend workarounds in gcc's cross-biarch.diff
It tweaks gcc to look in relevant multilib dirs and does so by
transforming gcc's build dir into the multilib libdir. This works fine
except for mips*. When gcc builds "n32", the libdir is "lib32" (already
covered in the patch). What wasn't covered is that gcc's "32" build
wants to use "libo32".