split the variable need_packages to built_packages
Before this commit need_packages was a set of source packages
initialized once and each built package was removed, but no package was
ever added after initialization.

Now built packages are added to the built_packages set and need_packages
can be augmented at any time. In particular,
automatically_cross_build_packages can add packages to it.

This split is necessary to differentiate packages that are not yet
needed from packages that are already built.