cross build check
In principle, we do not need check. It is mostly used in test suites and
we don't run test suites. Yet, libdebian-installer started to
Build-Depend on check. Until it makes that optional (#787044), we just
build check, because it is easier than patching Build-Depends. Once
libdebian-installer is fixed, check will simply no longer be needed and
thus no longer built.