cross build datefudge
It should not be necessary to cross build datefudge, because it cannot
be used in any sensible way for further cross build and is not part of
the build-essential installation set. We add it here, because it is
lists in gnutls28's Build-Depends without a <nocheck> profile. Just
building it is easier and more robust than adding yet another builddep_
hook.