avoid file conflict while upgrading gcc stages
It seems that differently staged builds of gcc-5 put different contents
into gcc-5-base. Now upgrading the stages of gcc-5-base can fail since,
because we both install gcc-5-base for multiple architectures and
upgrade it with the same version. Thus dpkg does not upgrade them in
lockstep and fails unpacking either architecture since the previous
(different) stage is still installed with the same version for the other
architecture.