patch glibc to drop dev package conflict
This patch is kinda wrong, because it the underlying file conflicts are
still in place. Also the build arch dev packages still carry conflicts.
The reason for dropping them already is that it allows us to see beyond
the declared conflict what other pieces need to be fixed to make those
dev packages truly coinstallable.