attempt to fix multilib glibc after multiarch includes
In a multilib setting, gcc does not add /usr/include/<triplet> to the
search path. Due to our multiarch include patch, all glibc headers
reside in such a path. The multilib packages were not updated in that
process and consequently cause failures now.

This commit only updates the mips64 multilib packages for mips.