carefully install stage3 gcc-N-base
When we upgrade gcc-N-base from one stage to the next, the version
doesn't change. Thus apt and dpkg believe them to be compatible, but
stage2 and stage3 use different patch sets and thus the READMEs differ.
We already remove the host arch gcc-N-base before upgrading, but we also
need to install the new build arch to avoid getting file conflicts when
the host arch one is unpacked first.

Reported-By: Steven Chamberlain <steven@pyro.eu.org>
Tested-By: Steven Chamberlain <steven@pyro.eu.org>