remove foreign libc-dev when building a supported cross compiler
The non-multiarch cross compiler build passes both the multiarch and the
sysroot paths to dpkg-shlibdeps in this order. When the foreign libc is
installed both as multiarch package and as a dpkg-crossed package, the
former is preferred and arch:all packages will end up depending on e.g.
libc6-ppc64, which are not satisfiable at all. When only the
dpkg-crossed packages are installed, satisfiable dependencies on e.g.
libc6-ppc64-powerpc-cross are generated instead.