try to work around diffoscope #813052
The working hypothesis is that diffoscope recurses over the working tree
containing the first package. This tree is big after building e.g.
glibc. So we copy the package into an empty directory now to confirm the
suspicion. The enhanced processing of dbgsym packages cannot be used
anyway, because DEB_BUILD_OPTIONS contains noddebs.