make downgrade to gcc-5 work again
It seems that apt is very unhappy to see packages with equal
name/arch/version combination but different content, so try hard to
avoid that. Also stop building gfortran (which is only used as a cross
tool) and remove duplicate libraries from the changes file.

It also seems that something breaks the detection of whether libc
provides -fstack-protector support. We already see this during the
stage3 cross compiler build, so some of our patches seem to break that.