cross build libidn2-0
This is a relatively new build dependency of curl. In order to build it,
bugs #856908 in gengetopt and #857675 in libidn2-0 need to be fixed.
Then it can be used with curl in place if libidn.

openssh-server is now annotated with <!nocheck> in curl, but curl gained
even more Build-Depends beyond libidn2-0 (e.g. libpsl), so we cannot
remove the builddep hook right away. For now, we keep building curl with
undersatisfied Build-Depends.