improve handling of gcc-N-base:native
gcc-N-base:native is present twice with the same version. Once in
unstable and once in the rebootstrap archive. This tends to confuse apt
and it sometimes thinks it should "downgrade" one to the other.

Instead of resolving this confusion by preferring one over the other,
try removing the excess version built with cross compilers and rely on
the archive version exclusively.