put rebootstrap first in sources.list
Now that we build gcc stage0 with fortran (parent commit), the
libgcc-6-dev package matches closely enough the unstable one that apt
cannot tell them apart. Thus it believes that the installed one is the
first matching one and that is unstable in this case. So it tries to
downgrade it every time due to the pinning. After reversing the order of
sources.list, apt thinks it comes from the rebootstrap list and is
happy.

Many thanks to David Kalnischkies and Julian Andres Klode for their
explanation of the background and diagnosis of the issue.