drop patch for building gcc without binutils-multiarch #804190
In gcc-7 and gcc-8, this patch is essentially obsoleted by
X-DH-Build-For-Type, which was added to gcc-7/7.2.0-19 and
gcc-8/8-20180110-1 via #879054. All that remains in #804190 is dropping
the binutils-multiarch dependency.

For gcc-6, the patch is still needed, but bootstrapping unstable with
gcc-6 no longer works at all anymore. That's just cruft removal.