fix building gzip for i386
gzip's configure has the following logic: If sizeof(time_t) < 8
(presently the case for all 32bit archs) and not cross_compiling
(presently the case for i386) and touch succeeds in storing a timestamp
for year 2106 (works on amd64), then it fails. This happens to trigger
for cross builds from amd64 to i386. Native builds for i386 succeed,
because touch fails. Cross builds for other 32bit architectures succeed,
because cross_compiling is yes there. Likely, this also fails for cross
builds from arm64 to armhf or from ppc64 to powerpc.